add deleteBookFromLS | addBookToUI as modules
