+++ modules/deleteBookFromUI.js
@@ -1,7 +1,8 @@
-let deleteBookFromUI = (el) =>{
-    if (el.classList.contains('remove-btn')) {
-        el.parentElement.remove();
-    }
-}
+const deleteBookFromUI = (el) => {
+  if (el.classList.contains('remove-btn')) {
+    el.parentElement.remove();
+  }
+};
 
-export {deleteBookFromUI}
+// eslint-disable-next-line import/prefer-default-export
+export { deleteBookFromUI };
